feat(xrpl/ws): errors,types and getSignedTx refactor, small naming fixes, improved SubmitTxAndWait function
