+++ xrpl/websocket/errors.go
@@ -4,7 +4,9 @@ import "errors"
 
 // Static errors
 var (
-	ErrMissingTxSignatureOrSigningPubKey = errors.New("transaction must have a TxSignature or SigningPubKey set")
+	ErrMissingTxSignatureOrSigningPubKey      = errors.New("transaction must have a TxSignature or SigningPubKey set")
+	ErrMissingLastLedgerSequenceInTransaction = errors.New("missing LastLedgerSequence in transaction")
+	ErrMissingWallet                          = errors.New("wallet must be provided when submitting an unsigned transaction")
 )
 
 // Dynamic errors
