+++ xrpl/websocket/types/client_options.go
@@ -1,0 +1,11 @@
+package types
+
+import (
+	"github.com/Peersyst/xrpl-go/xrpl/wallet"
+)
+
+type SubmitOptions struct {
+	Autofill bool
+	Wallet   *wallet.Wallet
+	FailHard bool
+}
